controllers: login, logout, auth

+++ studymate/base/views.py
@@ -1,7 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -10,6 +15,32 @@ from django.db.models import Q
 #     {'id': 2, 'name': 'Flask'},
 #     {'id': 3, 'name': 'DJango'}
 # ]
+
+
+def loginPage(request):
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist.')
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'User OR Password does not match.')
+
+    context = {}
+    return render(request, 'login_register.html', context)
+
+
+def logoutHandler(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -32,6 +63,7 @@ def room(request, pk):
     return render(request, 'room.html', context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -43,9 +75,14 @@ def create_room(request):
     return render(request, 'room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -56,8 +93,13 @@ def update_room(request, pk):
     return render(request, 'room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
